+++ spec/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.2].define(version: 2026_06_29_000002) do
+ActiveRecord::Schema[7.2].define(version: 2026_07_01_000001) do
   create_table "orders", force: :cascade do |t|
     t.datetime "cancelled_at"
     t.datetime "created_at", null: false
@@ -46,6 +46,8 @@ ActiveRecord::Schema[7.2].define(version: 2026_06_29_000002) do
     t.string "description"
     t.json "events", default: [], null: false
     t.json "metadata", default: {}
+    t.string "previous_secret"
+    t.datetime "previous_secret_expires_at"
     t.string "secret", null: false
     t.datetime "updated_at", null: false
     t.string "url", null: false